voice-patient-call-system/backend/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
azure-cognitiveservices-speech==1.42.0
azure-ai-textanalytics==5.3.0
python-dotenv==1.0.0
gunicorn==20.1.0